jira-ticket-summary/requirements.txt
requests==2.22.0
atlassian-python-api==1.16.0

